requirements.txt
httpx==0.27.0
selectolax==0.3.21
PyMuPDF==1.24.9
PyYAML==6.0.2
python-dateutil==2.9.0.post0

